Added president functions + tests
