+++ internal/clients/team3/utilfunctions.go
@@ -1,6 +1,7 @@
 package team3
 
 import (
+	"github.com/SOMAS2020/SOMAS2020/internal/common/gamestate"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/shared"
 )
 
@@ -38,4 +39,12 @@ func getAverage(lst []float64) float64 {
 	}
 
 	return (float64(total) / float64(len(lst)))
+}
+
+type mockServerReadHandle struct {
+	gameState gamestate.ClientGameState
+}
+
+func (m mockServerReadHandle) GetGameState() gamestate.ClientGameState {
+	return m.gameState
 }
